requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0.post1
python-magic==0.4.27
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0